# Conflicts:
#   backend/ycc/src/main/java/com/ai4ai/ycc/util/Scheduler.java

+++ backend/ycc/src/main/java/com/ai4ai/ycc/domain/profile/dto/request/AcceptLinkRequestDto.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Getter
 public class AcceptLinkRequestDto {
 
-    List<Integer> profiles;
+    List<Long> profiles;
 
 }
